ssl-proxies: Fix CRL refresh bug
The CertifcationRevocationList class caches CRL information to
reduce the cost of reloading them too often.

The caching logic contains a bug in which any call to the refresh
method will update the timestamp for the last reload even when
CRLs have not been reloaded. The consequence is that if refresh is
called sufficiently often, the CRLs will never be reloaded.