ssl-proxies: Add support for more RDN keywords
Java's X500Principal class only supports a limited number of
RDN keywords. Real world certificates make use of many more
keywords and bouncy castle and the EMI CANL library support
these. JGlobus 2 does not. This is a regression since JGlobus
1.x.

This patch adds additional mappings when mapping to and from
the X500Principal class. The mappings are copied from bouncy
castle 2.46 and CANL.

A unit test for a real world DN that we encountered on a
dCache installation has been added.