Update GlobusGSSException to avoid boring messages.
Quite often, a GlobusGSSException is used to wrap an existing Exception
where is adds no additional meaning.  Often this is due to catching
Exception but the signature of the method allows only GSSException to
be thrown, which is a bad code-smell.

The result of this problem is Exception messages with many chained
"Failure unspecified at GSS-API level" messages.

As a work-around, this patch skips over these semantically-null
Exception messages: if the message will be "Failure unspecified at
GSS-API level" then these are not printed, unless there is no cause
Exception.

The real solution would be to catch GSSExceptions and throw them in
situations where no additional context is provided.