Update GlobusGSSException to avoid boring messages
Quite often, a GlobusGSSException is used to wrap an existing
Exception where it adds no additional meaning.

The result of this problem is Exception messages with many chained
"Failure unspecified at GSS-API level" messages.

In addition, SSLEngine suffers from the same problem, throwing an
SSLException with a message that has no meaning and with a Throwable
(that cause the SSLException) containing the meaningful message.

As a work-around, this patch skips over these semantically-null
Exception messages.

This is really a work-around, but fixing the underlying problems will
take sufficiently long that such a work-around is necessary.