ssl-proxies: Add information about chain to CertPathValidationException
CertPathValidationException has fields to capture the CertPath and
certificate index that triggered the validation failure. This allows
an application to provide more information about the failure to the
user. Yet, JGlobus didn't fill in these fields of the exception,
meaning that the user would not be offered any information beyond
a simple error message saying that the certificate was rejected.

This patch adds the missing information. An application still has to
detect the exception in the exception cause chain and present it to
the user in some form.