Enable TLS 1.2 clients
SSLUtil fails to recognize TLS 1.2 packets. Regardless of whether
the server supports TLS 1.2, failure to recognize the header prevents
it from falling back to an earlier revision of the protocol.

The change enables forward compatibility. Based on my reading of the
TLS 1.2 RFC, I think the server should recognize versions higher than
it actually supports and reply with a lower version number. I assume
that as long as the TLS major version is unchanged, the header format
is compatible.